Move and rename deeplink APIs (#761)
* Move and rename deepLink APIs

* update test app

* revert test app component id changes

* rename test app component

* changelog
