+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 2.0.0-beta.3
+
+## Breaking Changes
+
+### Core APIs have been moved and renamed
+
+APIs previously under the `core` namespace, `executeDeepLink` and `shareDeepLink`, have moved.
+
+* `core.executeDeepLink` has renamed and moved to `app.openLink`
+* `core.shareDeepLink` has moved to `pages.shareDeepLink`
+
 # 2.0.0-beta.2
 
 ## Bug Fixes
